Helper function to mimic `.one()` in sqlalchemy
This allows us to raise our normal 404 (or whatever) errors we want when
doing a query that doesn't return any results, but it also allows us to
error out with a new 500 error if the results should only ever return
one item (due to a unique constraint on the database) and yet still
getting more then one item back from the results. That should obviously
never happen, if it did that indicates that a unique constraint is not
probably setup on the database or that I am using the `oneOrError`
method on something that I shouldn't be using it on.